backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
requests
python-dotenv
google-generativeai

